refactor FileOperations method
